My powers grow!!! Well, sort of....
A couple changes:
I removed the closing </input> tags, because input tags are
self-closing.  I also removed the id property and replaced it with name
properties, to try and make it fit better with what CodeCademy has had
me doing.
On the JS file, I added some variables to store the values from the text
inputs and then functions to pull them.
Within the createTheButton function, I call on those functions, and
nothing after that point in the function seems to work.