Fixes
1.  Nothing was happening because jQuery was not loaded.  See changes to
blocks.html for how to include.  I downloaded a copy of jQuery and added
it to the project.  You should also be able to link to it online, but
this is commented out in the html.

2.  Added 'px' to height and width, since that is how they are entered
in blocks.css.

3.  The div needs to be added to the document before you can modify its
properties.  I added a line break to separate different blocks, if they
are the same color.

4.  This still has one bug that I leave it to you to solve.  Right now
you create a div that has one ID.  You cannot have two elements with the
same ID on a page.  So, you need to find a way to make the ID's unique.
One idea is to implement a counter that gets appended to each ID when
you add it.  Afterwards the counter is incremented.

Great job!!

As a general note, I would try to learn to use the javascript debugger
in google chrome.  Hit F12, and click on Sources -> blocks.js.  It
immediately showed an error that '$' was not defined, letting me know
that jQuery was not loaded.